* changes:
  [qemu] Increase QEMUFile buffer size for faster IO
  More snapshot profiling logs - startup times
  Fix build for all platforms with SNAPSHOT_PROFILE
  [base] Add C wrapper for System's uptime getter
  [ui] Create extended window on first show, not in ctor
  [base] Faster loadBuffer() for stream serializing

+++ android/android-emugl/host/libs/libOpenglRender/RendererImpl.cpp
@@ -235,8 +235,8 @@ bool RendererImpl::load(android::base::Stream* stream) {
     cleanupRenderThreads();
     waitForProcessCleanup();
 #ifdef SNAPSHOT_PROFILE
-    printf("RenderThread cleanup time: %ld ms\n",
-            (android::base::System::get()->getUnixTimeUs()
+    printf("RenderThread cleanup time: %lld ms\n",
+            (long long)(android::base::System::get()->getUnixTimeUs()
             - startTime) / 1000);
 #endif
